[refactor/#62]:useCatalog hook 안에서 params 처리
