+++ src/features/Catalog/components/Main/ProductGrid.tsx
@@ -1,27 +1,12 @@
 import styled from 'styled-components';
 import Pagination from '@/components/Pagination/Pagination';
-import { useLocation, useNavigate } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import ProductCard from './ProductCard';
 import { useCatalog } from '../../hooks/useCatalog';
 
 function ProductGrid() {
-  const location = useLocation();
-  const searchParams = new URLSearchParams(location.search);
   const navigate = useNavigate();
-
-  const page = searchParams.get('page') || 1;
-  const category = searchParams.get('category');
-  const search = searchParams.get('search');
-  const minAlcohol = searchParams.get('minAlcohol');
-  const maxAlcohol = searchParams.get('maxAlcohol');
-
-  const { catalogData } = useCatalog({
-    page: Number(page),
-    category: category ? +category : undefined,
-    search: search || undefined,
-    minAlcohol: minAlcohol ? +minAlcohol : undefined,
-    maxAlcohol: maxAlcohol ? +maxAlcohol : undefined,
-  });
+  const { catalogData } = useCatalog();
 
   return (
     <Container>
