+++ src/features/Catalog/hooks/useCatalog.ts
@@ -1,9 +1,26 @@
 import { useQuery } from '@tanstack/react-query';
+import { useLocation } from 'react-router-dom';
 import { getCatalog } from '../api/catalog.api';
 import { CatalogParams } from '../types/CatalogParams';
 import { CatalogItem } from '../types/catalog';
 
-export const useCatalog = (params: CatalogParams) => {
+export const useCatalog = () => {
+  const location = useLocation();
+  const searchParams = new URLSearchParams(location.search);
+
+  const page = searchParams.get('page') || '1';
+  const category = searchParams.get('category');
+  const search = searchParams.get('search');
+  const minAlcohol = searchParams.get('minAlcohol');
+  const maxAlcohol = searchParams.get('maxAlcohol');
+
+  const params: CatalogParams = {
+    page: Number(page),
+    category: category ? +category : undefined,
+    search: search || undefined,
+    minAlcohol: minAlcohol ? +minAlcohol : undefined,
+    maxAlcohol: maxAlcohol ? +maxAlcohol : undefined,
+  };
   const {
     data: catalogData,
     error,
